+++ client/js/components/statement/StatementExportModal.vue
@@ -27,12 +27,11 @@
         <legend
           class="o-form__label text-base"
           v-text="Translator.trans('export.type')" />
-        <div class="flex flex-row mb-5 mt-1 gap-3">
+        <div class="grid grid-cols-3 mt-2 mb-5 gap-x-2 gap-y-5">
           <dp-radio
             v-for="(exportType, key) in exportTypes"
             :key="key"
             :id="key"
-            class="max-w-[70%]"
             :data-cy="`exportType:${key}`"
             :label="{
               hint: active === key ? exportType.hint : '',
@@ -44,7 +43,7 @@
         </div>
       </fieldset>
 
-      <fieldset v-if="['docx', 'zip'].includes(this.active)">
+      <fieldset v-if="['docx_normal', 'docx_censored', 'zip_normal', 'zip_censored'].includes(this.active)">
         <legend
           id="docxColumnTitles"
           class="o-form__label text-base float-left mr-1"
@@ -134,7 +133,7 @@ export default {
 
   data () {
     return {
-      active: 'docx',
+      active: 'docx_normal',
       docxColumns: {
         col1: {
           width: 'col-span-1',
@@ -156,23 +155,37 @@ export default {
         }
       },
       exportTypes: {
-        docx: {
+        docx_normal: {
           label: 'export.docx',
           hint: '',
           exportPath: 'dplan_statement_segments_export',
           dataCy: 'exportModal:export:docx'
         },
-        zip: {
+        zip_normal: {
           label: 'export.zip',
           hint: '',
           exportPath: 'dplan_statement_segments_export_packaged',
           dataCy: 'exportModal:export:zip'
         },
-        xlsx: {
+        xlsx_normal: {
           label: 'export.xlsx',
           hint: Translator.trans('export.xlsx.hint'),
           exportPath: 'dplan_statement_xls_export',
           dataCy: 'exportModal:export:xlsx'
+        },
+        docx_censored: {
+          label: 'export.docx.censored',
+          hint: '',
+          exportPath: 'dplan_statement_segments_export',
+          dataCy: 'exportModal:export:docx',
+          censor: true
+        },
+        zip_censored: {
+          label: 'export.zip.censored',
+          hint: '',
+          exportPath: 'dplan_statement_segments_export_packaged',
+          dataCy: 'exportModal:export:zip',
+          censor: true
         }
       },
       fileName: '',
@@ -216,7 +229,7 @@ export default {
 
     handleExport () {
       const columnTitles = {}
-      const shouldConfirm = ['docx', 'zip'].includes(this.active)
+      const shouldConfirm = /^(docx|zip)_/.test(this.active)
 
       Object.keys(this.docxColumns).forEach(key => {
         const columnTitle = this.docxColumns[key].title
@@ -233,9 +246,10 @@ export default {
 
       this.$emit('export', {
         route: this.isSingleStatementExport ? this.singleStatementExportPath : this.exportTypes[this.active].exportPath,
-        docxHeaders: ['docx', 'zip'].includes(this.active) ? columnTitles : null,
+        docxHeaders: ['docx_normal', 'docx_censored', 'zip_normal', 'zip_censored'].includes(this.active) ? columnTitles : null,
         fileNameTemplate: this.fileName || null,
-        shouldConfirm
+        shouldConfirm,
+        censorParameter: this.exportTypes[this.active].censor || false
       })
       this.closeModal()
     },
@@ -246,7 +260,7 @@ export default {
     },
 
     setInitialValues () {
-      this.active = 'docx'
+      this.active = 'docx_normal'
 
       Object.keys(this.docxColumns).forEach(key => {
         const storageKey = `exportModal:docxCol:${key}`