+++ client/js/components/tags/TagsList.vue
@@ -2,7 +2,9 @@
   <div>
     <tags-list-header />
 
-    <tags-create-form :procedure-id="procedureId" />
+    <tags-create-form
+      :is-master-procedure="isMasterProcedure"
+      :procedure-id="procedureId" />
 
     <dp-tree-list
       v-if="transformedCategories"
@@ -44,7 +46,7 @@
           :is-in-edit-state="isInEditState"
           :procedure-id="procedureId"
           type="TagTopic"
-          @abort="abort"
+          @abort="closeEditForm"
           @delete="deleteItem"
           @edit="setEditState"
           @save="save" />
@@ -55,7 +57,7 @@
           :is-in-edit-state="isInEditState"
           :procedure-id="procedureId"
           type="Tag"
-          @abort="abort"
+          @abort="closeEditForm"
           @delete="deleteItem"
           @edit="setEditState"
           @save="save" />
@@ -73,49 +75,43 @@
 <script>
 import {
   checkResponse,
-  DpButton,
-  DpIcon,
-  DpInput,
   DpLoading,
-  DpModal,
   dpRpc,
-  DpTreeList,
-  DpUpload
+  DpTreeList
 } from '@demos-europe/demosplan-ui'
 import { mapActions, mapMutations, mapState } from 'vuex'
 import AddonWrapper from '@DpJs/components/addon/AddonWrapper'
+import TagListEditForm from './TagListEditForm'
 import TagsCreateForm from './TagsCreateForm'
 import TagsImportForm from './TagsImportForm'
-import TagListBulkControls from './TagListBulkControls'
-import TagListEditForm from './TagListEditForm'
 import TagsListHeader from './TagsListHeader'
 export default {
   name: 'TagsList',
 
   components: {
     AddonWrapper,
-    DpButton,
-    DpIcon,
-    DpInput,
     DpLoading,
-    DpModal,
-    DpUpload,
     DpTreeList,
     TagsCreateForm,
     TagsImportForm,
-    TagListBulkControls,
     TagListEditForm,
     TagsListHeader
   },
 
   props: {
+    isMasterProcedure: {
+      type: Boolean,
+      required: false,
+      default: false
+    },
+
     procedureId: {
       type: String,
       required: true
     }
   },
 
-  data() {
+  data () {
     return {
       dataIsRequested: false,
       isInEditState: ''
@@ -177,7 +173,7 @@ export default {
       saveTagTopic: 'save'
     }),
 
-    abort () {
+    closeEditForm () {
       this.isInEditState = ''
     },
 
@@ -186,17 +182,21 @@ export default {
         id: parentTopic.id,
         type: 'TagTopic',
         attributes: parentTopic.attributes,
-        relationships: {
-          tags: {
-            data: [
-              ...parentTopic.relationships.tags.data,
-              {
-                id: tagId,
-                type: 'Tag'
+        relationships: parentTopic.relationships
+          ? {
+              ...parentTopic.relationships,
+              tags: {
+                data: parentTopic.relationships.tags.data.concat({
+                  type: 'Tag',
+                  id: tagId
+                })
               }
-            ]
-          }
-        }
+            }
+          : {
+              tags: {
+                data: [{ type: 'Tag', id: tagId }]
+              }
+            }
       })
 
       this.saveTagTopic(parentTopic.id)
@@ -209,11 +209,11 @@ export default {
         return
       }
 
-      const hasNewParent = !this.TagTopic[parentId].relationships.tags.data.find(tag => tag.id === elementId)
+      const hasNewParent = !this.TagTopic[parentId].relationships?.tags.data.find(tag => tag.id === elementId)
 
       if (hasNewParent) {
         const parentTopic = { ...this.TagTopic[parentId] }
-        const oldParent = Object.values(this.TagTopic).find(topic => topic.relationships.tags.data.find(tag => tag.id === elementId))
+        const oldParent = Object.values(this.TagTopic).find(topic => topic.relationships?.tags.data.find(tag => tag.id === elementId))
 
         this.addTagToNewTopic(parentTopic, elementId)
         this.removeTagFromOldTopic(oldParent, elementId)
@@ -277,22 +277,25 @@ export default {
       this.saveTagTopic(oldParent.id)
     },
 
-    save ({ id, attributes, type }) {
-      if (id === '') {
+    save ({ id, attributes, type, isTitleChanged }) {
+      if (!id || !isTitleChanged) {
+        this.closeEditForm()
+
         return
       }
 
-      console.log('save', { id, attributes, type })
+      const updateMethod = `update${type}`
+      const saveMethod = `save${type}`
 
-      this[`update${type}`]({
+      this[updateMethod]({
         attributes,
         id,
-        relationships: this[type][id].relationships,
+        relationships: this[type][id]?.relationships,
         type
       })
-      this[`save${type}`](id)
+      this[saveMethod](id)
         .then(() => {
-          this.isInEditState = ''
+          this.closeEditForm()
         })
     },