+++ demosplan/DemosPlanCoreBundle/ResourceTypes/TagTopicResourceType.php
@@ -141,9 +141,17 @@ final class TagTopicResourceType extends DplanResourceType
             )
         );
 
-        $configBuilder->procedure
-            ->setRelationshipType($this->resourceTypeStore->getProcedureResourceType())
-            ->setReadableByPath()->setSortable()->setFilterable()->initializable();
+        if ($this->currentProcedureService->getProcedure()->getMaster()) {
+            $configBuilder->procedure
+                ->setRelationshipType($this->resourceTypeStore->getProcedureTemplateResourceType())
+                ->setReadableByPath()->setSortable()->setFilterable()->initializable();
+        }
+        if (!$this->currentProcedureService->getProcedure()->getMaster()) {
+            $configBuilder->procedure
+                ->setRelationshipType($this->resourceTypeStore->getProcedureResourceType())
+                ->setReadableByPath()->setSortable()->setFilterable()->initializable();
+        }
+
         $configBuilder->addConstructorBehavior(
             new FixedConstructorBehavior(
                 Paths::tagTopic()->procedure->getAsNamesInDotNotation(),