+++ CHANGELOG.md
@@ -5,18 +5,32 @@
 - **Patch Version**: Incremented for bug fixes.
 
 ## UNRELEASED
+- Enable to send Statement final notice using RpcRequest and Vue.js
+- Add the possibility to export synopsis without personal data.
+
+## v2.26.2 (2025-03-14)
+- Create report entries on create, update, deletion of an element, paragraph , singleDocument, mapDrawing or mapDrawing-explanation
+
+## v2.27.0 (2025-03-12)
+- Enable to send Statement final notice using RpcRequest and Vue.js
+- Add the possibility to export synopsis without personal data.
+
+## v2.26.0 (2025-02-25)
 - Allow flag on external links to indicate this URL should only be shown for user(roles) with a specific permission
 - Extend safelist for purge css to include all plyr classes
 
 ### Features
 - Institution tag management: Add search field and filters to institution list
 - Procedure basic settings: Move procedure location up under the "internal" section
+- TagsListEditForm: add a confirmation message before deleting the Tag or Topic
 
 ### Further changes
 - Segments list: Use DpSearchField for custom search
 - DpInlineNotification: Set margin from outside the component (instead of inside)
 - Rework Tags-Lists: Now works with APi2.0 and is extendable by addons.
-
+- Enable the DELETE method for the AdministratableUser resource type
+- TagsList: trigger an update request only when the title is modified
+- DpCreateTag: Ensure the new tag retains its relationship to the Topic after creation
 
 ## 2.20.0